[FEAT] separate RingPainter from MiniCalendar
