+++ lib/app/views/todo/widget/peep_mini_calendar.dart
@@ -1,65 +1,14 @@
-import 'dart:math';
-import 'dart:developer' as dev;
-
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:get/get.dart';
 import 'package:intl/intl.dart';
 import 'package:peep_todo_flutter/app/controllers/mini_calendar_controller.dart';
 import 'package:peep_todo_flutter/app/controllers/page/scheduled_todo_controller.dart';
 import 'package:peep_todo_flutter/app/theme/text_style.dart';
-
+import 'package:peep_todo_flutter/app/views/common/painter/ring_painter.dart';
 import 'package:table_calendar/table_calendar.dart';
-
-import '../../../data/model/category_model.dart';
 import '../../../theme/app_values.dart';
 import '../../../theme/palette.dart';
-
-class ringPainter extends CustomPainter {
-  final List<CategoryModel> categoryList;
-  final List<double> itemCounts;
-
-  ringPainter({required this.itemCounts, required this.categoryList});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    const double total = 1 / 2;
-    const double startAngle = -pi / 2;
-
-    double currentAngle = startAngle;
-
-    if(itemCounts.isEmpty) return;
-
-    for (int i = 0; i < categoryList.length; i++) {
-      final sweepAngle = ((itemCounts[i] / total) * pi); // 아이템 수에 따른 각도
-      final rect = Rect.fromCircle(
-        center: Offset(size.width / 2, size.height / 2),
-        radius: size.width / 2,
-      );
-
-      final paint = Paint()
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = 5.0
-        ..color = categoryList[i].color.withOpacity(AppValues.halfOpacity); // 아이템 순위에 해당하는 색상 사용
-
-      canvas.drawArc(rect, currentAngle, sweepAngle, false, paint);
-
-      currentAngle += sweepAngle;
-    }
-  }
-
-  @override
-  bool shouldRepaint(CustomPainter oldDelegate) {
-    if (oldDelegate is ringPainter) {
-      return !listEquals(itemCounts, oldDelegate.itemCounts) ||
-          !listEquals(categoryList, oldDelegate.categoryList);
-    }
-    return true;
-  }
-}
 
 class PeepMiniCalendar extends StatelessWidget {
   final ScheduledTodoController controller;
@@ -222,7 +171,6 @@ class PeepMiniCalendar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-
     return Obx(() => Container(
           decoration: BoxDecoration(
               color: Palette.peepWhite,
@@ -258,11 +206,14 @@ class PeepMiniCalendar extends StatelessWidget {
                             ),
                           ),
                         ),
-                        Obx(()=>
-                          CustomPaint(
+                        Obx(
+                          () => CustomPaint(
                             size: Size(32.w, 32.h), // CustomPaint의 크기 고정
-                            painter:
-                                ringPainter(itemCounts: controller.calendarItemCounts[DateFormat('yyyyMMdd').format(day)] ?? [], categoryList: controller.getCategoryList()),
+                            painter: RingPainter(
+                                itemCounts: controller.calendarItemCounts[
+                                        DateFormat('yyyyMMdd').format(day)] ??
+                                    [],
+                                categoryList: controller.getCategoryList()),
                           ),
                         ),
                         Center(
@@ -282,11 +233,14 @@ class PeepMiniCalendar extends StatelessWidget {
                     child: Stack(
                       alignment: Alignment.center,
                       children: [
-                        Obx(()=>
-                          CustomPaint(
+                        Obx(
+                          () => CustomPaint(
                             size: Size(32.w, 32.h),
-                            painter:
-                            ringPainter(itemCounts: controller.calendarItemCounts[DateFormat('yyyyMMdd').format(day)] ?? [], categoryList: controller.getCategoryList()),
+                            painter: RingPainter(
+                                itemCounts: controller.calendarItemCounts[
+                                        DateFormat('yyyyMMdd').format(day)] ??
+                                    [],
+                                categoryList: controller.getCategoryList()),
                           ),
                         ),
                         Text(
